+++ ubo_app/store/__init__.py
@@ -19,6 +19,7 @@ from ubo_app.store.services.camera import CameraAction, CameraEvent
 from ubo_app.store.services.ip import IpAction, IpEvent, IpState
 from ubo_app.store.services.keypad import KeypadEvent
 from ubo_app.store.services.notifications import NotificationsAction, NotificationsState
+from ubo_app.store.services.sensors import SensorsAction, SensorsState
 from ubo_app.store.services.sound import SoundAction, SoundState
 from ubo_app.store.services.wifi import WiFiAction, WiFiEvent, WiFiState
 from ubo_app.store.status_icons import StatusIconsAction, StatusIconsState
@@ -31,6 +32,7 @@ class RootState(BaseCombineReducerState):
     main: MainState
     status_icons: StatusIconsState
     update_manager: UpdateManagerState
+    sensors: SensorsState
     sound: SoundState
     wifi: WiFiState
     ip: IpState
@@ -42,6 +44,7 @@ ActionType = (
     | StatusIconsAction
     | UpdateManagerAction
     | MainAction
+    | SensorsAction
     | SoundAction
     | CameraAction
     | WiFiAction